feat(left-sidebar)- implement the sidebar and fix the layout and refactor the code
