+++ __mocks__/mockSongsData.js
@@ -1,236 +1,110 @@
 export const mockMusicData = [
   {
+    id: 1,
+    trackNumber: 1,
+    title: "Havana",
+    artist: "Annie Lennox",
+    duration: "20:24",
     cover: "https://i.scdn.co/image/ab67616d0000b2732a038d3bf875d23e4aeaa84e",
-    songName: "BIRDS OF A FEATHER",
-    singers: "Billie Eilish",
-    length: "3:30",
+    color: "bg-gradient-to-b from-pink-400 via-purple-500 to-blue-500",
   },
   {
+    id: 2,
+    trackNumber: 2,
+    title: "BIRDS OF A FEATHER",
+    artist: "Billie Eilish",
+    duration: "3:30",
+    cover: "https://i.scdn.co/image/ab67616d0000b2732a038d3bf875d23e4aeaa84e",
+    color: "bg-gradient-to-b from-green-400 to-blue-500",
+  },
+  {
+    id: 3,
+    trackNumber: 3,
+    title: "Galway Girl",
+    artist: "Ed Sheeran",
+    duration: "2:49",
     cover: "https://i.scdn.co/image/ab67616d0000b273ba5db46f4b838ef6027e6f96",
-    songName: "Galway Girl",
-    singers: "Ed Sheeran",
-    length: "2:49",
+    color: "bg-gradient-to-b from-orange-400 to-red-500",
   },
   {
-    cover: "https://i.scdn.co/image/ab67616d00001e02472fbc1d5743c7d3c75b9ec0",
-    songName: "Just the two of us",
-    singers: "Grover Washington, Bill Withers",
-    length: "7:19",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e02727d531901c07a499498c544",
-    songName: "Zara sa",
-    singers: "KK",
-    length: "5:03",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e02cc01790e9775b822640c14c8",
-    songName: "O meri jaan",
-    singers: "KK",
-    length: "4:20",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00004851417cc363d7286a88489dee3c",
-    songName: "The Girl from Ipanema",
-    singers: "Stan Getz",
-    length: "5:19",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e02238513a0da74c837b2eeb8d7",
-    songName: "Kya Mujhe Pyaar Hai",
-    singers: "KK",
-    length: "4:26",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e02a10a214d4d4935347e8079b4",
-    songName: "Can't take My Eyes off you",
-    singers: "Frankie Valli",
-    length: "3:23",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e028db55b7cb540f7ab7c344d1b",
-    songName: "Two in Love",
-    singers: "Nat King Cole",
-    length: "3:43",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b2738daeb3002870ebba8e8d1387",
-    songName: "L-O-V-E",
-    singers: "Nat King Cole",
-    length: "2:35",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e02ff0dae802acb38075786b58c",
-    songName: "Fly me to the moon",
-    singers: "Frank Sinatara",
-    length: "2:27",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e021da0a37c5bad81580d49ef05",
-    songName: "Tujhe Bhula Diya",
-    singers: "Mohit Chauhan",
-    length: "4:40",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e02668e3aca3167e6e569a9aa20",
-    songName: "Kill Bill",
-    singers: "SZA",
-    length: "2:33",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e02a991995542d50a691b9ae5be",
-    songName: "Antifragile",
-    singers: "LE SSERAFIM",
-    length: "3:04",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e02b55d26c578e30129b0a7e86e",
-    songName: "Remedi",
-    singers: "Tulus",
-    length: "4:07",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b27356ac7b86e090f307e218e9c8",
-    songName: "Espresso",
-    singers: "Sabrina Carpenter",
-    length: "2:55",
-  },
-  {
+    id: 4,
+    trackNumber: 4,
+    title: "Anti-Hero",
+    artist: "Taylor Swift",
+    duration: "3:20",
     cover: "https://i.scdn.co/image/ab67616d0000b273bb54dde68cd23e2a268ae0f5",
-    songName: "Anti-Hero",
-    singers: "Taylor Swift",
-    length: "3:20",
+    color: "bg-gradient-to-b from-purple-400 to-pink-500",
   },
   {
-    cover: "https://i.scdn.co/image/ab67616d00001e0254e544672baa16145d67612b",
-    songName: "Kun Faya Kun",
-    singers: "Mohit Chauhan",
-    length: "7:50",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b2732a038d3bf875d23e4aeaa84e",
-    songName: "LUNCH",
-    singers: "Billie Eilish",
-    length: "2:59",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b273f7b7174bef6f3fbfda3a0bb7",
-    songName: "vampire",
-    singers: "Olivia Rodrigo",
-    length: "3:39",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e0252fe6875028c892308ffc2f7",
-    songName: "Tum Se Hi",
-    singers: "Mohit Chauhan",
-    length: "5:21",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e02bcb8ec035de7b2dda0f74660",
-    songName: "Dil Ibaadat",
-    singers: "KK",
-    length: "5:29",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d00001e02ae40468931087f78919b86ce",
-    songName: "Leave Before You Love Me",
-    singers: "Jonas Brothers",
-    length: "3:24",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b273cd222052a2594be29a6616b5",
-    songName: "Unholy",
-    singers: "Sam Smith, Kim Petras",
-    length: "2:36",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b27328933b808bfb4cbbd0385400",
-    songName: "Industry Baby",
-    singers: "Lil Nas X, Jack Harlow",
-    length: "3:32",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b273a935e4689f15953311772cc4",
-    songName: "Fortnight",
-    singers: "Taylor Swift, Post Malone",
-    length: "3:48",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b273b7e976d2b35c767f9012cb72",
-    songName: "Golden Hour",
-    singers: "JVKE",
-    length: "3:29",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b273a91c10fe9472d9bd89802e5a",
-    songName: "Stick Season",
-    singers: "Noah Kahan",
-    length: "3:01",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b273a91c10fe9472d9bd89802e5a",
-    songName: "Dial Drunk",
-    singers: "Noah Kahan",
-    length: "2:51",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b273ae395b47b186c2bc8c458e0f",
-    songName: "Creepin'",
-    singers: "Metro Boomin, The Weeknd, 21 Savage",
-    length: "3:41",
-  },
-  {
-    cover: "https://i.scdn.co/image/ab67616d0000b273b7e976d2b35c767f9012cb72",
-    songName: "this is what falling in love feels like",
-    singers: "JVKE",
-    length: "2:11",
-  },
-  {
+    id: 5,
+    trackNumber: 5,
+    title: "Blinding Lights",
+    artist: "The Weeknd",
+    duration: "3:20",
     cover: "https://i.scdn.co/image/ab67616d00001e02ff9ca10b55ce82ae553c8228",
-    songName: "Blinding Lights",
-    singers: "The Weeknd",
-    length: "3:20",
+    color: "bg-gradient-to-b from-yellow-400 to-orange-500",
   },
   {
-    cover: "https://i.scdn.co/image/ab67616d00001e0295f754318336a07e85ec59bc",
-    songName: "Flowers",
-    singers: "Miley Cyrus",
-    length: "3:20",
+    id: 6,
+    trackNumber: 6,
+    title: "Golden Hour",
+    artist: "JVKE",
+    duration: "3:29",
+    cover: "https://i.scdn.co/image/ab67616d0000b273b7e976d2b35c767f9012cb72",
+    color: "bg-gradient-to-b from-blue-400 to-purple-500",
   },
   {
-    cover: "https://i.scdn.co/image/ab67616d00001e02668e3aca3167e6e569a9aa20",
-    songName: "Kill Bill",
-    singers: "SZA",
-    length: "2:33",
+    id: 7,
+    trackNumber: 7,
+    title: "Shape of You",
+    artist: "Ed Sheeran",
+    duration: "3:53",
+    cover: "https://i.scdn.co/image/ab67616d0000b273ba5db46f4b838ef6027e6f96",
+    color: "bg-gradient-to-b from-red-400 to-yellow-500",
   },
   {
-    cover: "https://i.scdn.co/image/ab67616d00001e02a991995542d50a691b9ae5be",
-    songName: "As It Was",
-    singers: "Harry Styles",
-    length: "2:47",
+    id: 8,
+    trackNumber: 8,
+    title: "bad guy",
+    artist: "Billie Eilish",
+    duration: "3:14",
+    cover: "https://i.scdn.co/image/ab67616d0000b2732a038d3bf875d23e4aeaa84e",
+    color: "bg-gradient-to-b from-green-300 to-yellow-400",
   },
   {
-    cover: "https://i.scdn.co/image/ab67616d00001e02b55d26c578e30129b0a7e86e",
-    songName: "Levitating",
-    singers: "Dua Lipa",
-    length: "3:23",
+    id: 9,
+    trackNumber: 9,
+    title: "Save Your Tears",
+    artist: "The Weeknd",
+    duration: "3:35",
+    cover: "https://i.scdn.co/image/ab67616d00001e02ff9ca10b55ce82ae553c8228",
+    color: "bg-gradient-to-b from-blue-400 to-pink-400",
   },
   {
-    cover: "https://i.scdn.co/image/ab67616d00001e02e6f407c7f3a0ec98845e4431",
-    songName: "Dynamite",
-    singers: "BTS",
-    length: "3:19",
+    id: 10,
+    trackNumber: 10,
+    title: "As It Was",
+    artist: "Harry Styles",
+    duration: "2:47",
+    cover: "https://i.scdn.co/image/ab67616d0000b273b7e976d2b35c767f9012cb72",
+    color: "bg-gradient-to-b from-pink-300 to-blue-300",
   },
   {
-    cover: "https://i.scdn.co/image/ab67616d00001e02b1c4b76e23414c9f20242268",
-    songName: "drivers license",
-    singers: "Olivia Rodrigo",
-    length: "4:02",
+    id: 11,
+    trackNumber: 11,
+    title: "Flowers",
+    artist: "Miley Cyrus",
+    duration: "3:21",
+    cover: "https://i.scdn.co/image/ab67616d0000b273b7e976d2b35c767f9012cb72",
+    color: "bg-gradient-to-b from-yellow-300 to-pink-400",
   },
   {
-    cover: "https://i.scdn.co/image/ab67616d00001e02b1c4b76e23414c9f20242268",
-    songName: "Sunflower",
-    singers: "Post Malone, Swae Lee",
-    length: "2:38",
+    id: 12,
+    trackNumber: 12,
+    title: "Dance The Night",
+    artist: "Dua Lipa",
+    duration: "2:56",
+    cover: "https://i.scdn.co/image/ab67616d0000b273b7e976d2b35c767f9012cb72",
+    color: "bg-gradient-to-b from-purple-300 to-pink-300",
   },
 ];
