+++ components/dashboard/index.tsx
@@ -5,7 +5,7 @@ import { DashboardSongs } from "./dashboard-songs";
 
 export function Dashboard() {
   return (
-    <div className="p-4 flex flex-col gap-4">
+    <div className="p-4 grid grid-rows-12 gap-4 h-full">
       <Header />
       <HeroSection />
       <DashboardSongs />
