+++ components/sidebar/Library.tsx
@@ -125,7 +125,7 @@ const PlaylistList = ({ items }: { items: typeof playlists }) => (
 
 export function Library() {
   return (
-    <Tabs defaultValue="playlists" className="max-w-xs w-full">
+    <Tabs defaultValue="playlists" className="max-w-xs w-full ">
       <TabsList className="w-full grid grid-cols-3 gap-2 px-2">
         <TabsTrigger
           className="text-xs bg-white/20 rounded-2xl cursor-pointer hover:bg-white/40 transition-all duration-300"
